Moved Tags Composables to their own file
